python/libraries/requirements.txt
modbus_tk==1.1.2
pyserial==3.5
